Rework NetworkBus
This reworks NetworkBus in an attempt to address asaskevich/EventBus/#28

Previously, remote subscriptions were handled via regular handlers in
the bus, which would publish the event in the remote bus by invoking its
Publish() method over RPC.

If the bus originally publishing the event was also subscribed to the
same topic on the remote bus, then its "remote" handler would be called
in the remote bus, calling back into the original caller, and
deadlocking (or calling back and forth forever).

To remedy this, NetworkBus is split into two separate busses - one only
holds handlers which correspond to remote subscribers, and one only
handling "local" subscribers. When an event is published (locally), it
gets broadcast to both local and remote subscribers. However, when an
event is received over the network, it is _only_ published to local
subscribers. This prevents the call back to the remote originator.

The downside to this approach is that:
 a) Two busses are needed (which uses resources)
 b) Events received over the network are not propagated to other
    remote subscribers.
    i.e. If BusA subscribes to "topic1" in BusB, and BusB subscribes to
    "topic1" in BusC, then if someone publishes "topic1" in BusC, then
    it gets received in BusB, but _not_ in BusA. BusA would need to
    subscribe directly to BusC in order to receive the event.

*NOTE*: This breaks compatibility by renaming Client.Subscribe[Once] to
Client.Subscribe[Once]Remote. This allows NetworkBus to expose the
"normal" Bus interface, but might be unacceptable.